# api/swagger.yaml
openapi: 3.0.3
info:
  title: Viral Content Scraper API
  description: |
    # 🚀 API de Scraping Inteligente de Conteúdo Viral
    
    Uma API completa para coleta, análise e geração de insights sobre conteúdo viral em múltiplas plataformas sociais.
    
    ## 🎯 Funcionalidades Principais
    
    - **Scraping Inteligente**: Coleta automatizada de conteúdo viral
    - **Análise com IA**: Processamento avançado com GPT-4 Vision
    - **Templates Virais**: Extração e adaptação de padrões de sucesso
    - **Análise de Perfis**: Análise completa de perfis Instagram
    - **Tendências**: Identificação de padrões virais em tempo real
    - **Administração**: Gerenciamento completo do sistema
    - **Webhooks**: Notificações em tempo real
    
    ## 🔐 Autenticação
    
    A API utiliza autenticação JWT Bearer Token:
    ```
    Authorization: Bearer <seu_token_jwt>
    ```
    
    ## 📊 Rate Limiting
    
    - **Usuário Padrão**: 1000 requests/hora, 100 requests/minuto
    - **Usuário Premium**: 5000 requests/hora, 500 requests/minuto
    - **Usuário Business**: 15000 requests/hora, 1500 requests/minuto
    
    ## 🌐 Base URL
    
    - **Produção**: `https://api.viralscraper.com/api/v1`
    - **Desenvolvimento**: `http://localhost:5000/api/v1`
    
  version: 1.0.0
  contact:
    name: Manus AI
    email: support@viralscraper.com
    url: https://viralscraper.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /templates:
    get:
      tags: [Templates]
      summary: Listar templates
      description: Busca templates com filtros avançados
      parameters:
        - name: content_type
          in: query
          schema:
            type: string
            enum: [carousel, reel, post, story]
        - name: min_viral_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: platform
          in: query
          schema:
            type: string
            enum: [instagram, tiktok, linkedin, youtube]
        - name: search_tags
          in: query
          schema:
            type: string
            description: Tags separadas por vírgula
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [viral_score, created_at, template_quality_score]
            default: viral_score
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [desc, asc]
            default: desc
      responses:
        '200':
          description: Lista de templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplatesListResponse'

    post:
      tags: [Templates]
      summary: Criar template personalizado
      description: Cria um novo template baseado em especificações
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticação de usuários
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT com privilégios de administrador
  schemas:
...
    TemplateResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            template_id:
              type: string
            template_name:
              type: string
            content_type:
              type: string
            viral_score:
              type: integer
            visual_structure:
              type: object
            content_formula:
              type: object
            performance_metrics:
              type: object
            extraction_metadata:
              type: object
        timestamp:
          type: string
          format: date-time
...
    CreateTemplateRequest:
      type: object
      properties:
        template_name:
          type: string
        content_type:
          type: string
          enum: [carousel, reel, post, story]
        visual_structure:
          type: object
        content_formula:
          type: object
        target_platforms:
          type: array
          items:
            type: string
        search_tags:
          type: array
          items:
            type: string
      required: [template_name, content_type]

    # ERROR RESPONSES
...
    TemplatesListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            templates:
              type: array
              items:
                $ref: '#/components/schemas/TemplateResponse'
            pagination:
              type: object
              properties:
                total:
                  type: integer
                limit:
                  type: integer
                offset:
                  type: integer
                has_more:
                  type: boolean
...